test: comp request on leave period boundary
