+++ hrms/tests/test_utils.py
@@ -69,6 +69,21 @@ def get_first_day_for_prev_month():
 	return prev_month_first
 
 
+def add_date_to_holiday_list(date: str, holiday_list: str) -> None:
+	if frappe.db.exists("Holiday", {"parent": holiday_list, "holiday_date": date}):
+		return
+
+	holiday_list = frappe.get_doc("Holiday List", holiday_list)
+	holiday_list.append(
+		"holidays",
+		{
+			"holiday_date": date,
+			"description": "test",
+		},
+	)
+	holiday_list.save()
+
+
 def create_company(name: str = "_Test Company"):
 	if frappe.db.exists("Company", name):
 		return frappe.get_doc("Company", name)
